:recycle: refactors to simplify things
